## Summary - Fix Junit test ## Trello Card https://trello.com/c/uKZUWHek/28-junit-testing
